Add folders, replace classes and add more interfaces
