+++ SOLIDHomework.Core/Calculators/IDiscountCalculator.cs
@@ -1,0 +1,7 @@
+﻿namespace SOLIDHomework.Core.Calculators
+{
+    public interface IDiscountCalculator
+    {
+        decimal CalculateUnitDiscount(OrderItem orderItem);
+    }
+}
