+++ SOLIDHomework.Core/Calculators/IItemCalculator.cs
@@ -1,0 +1,7 @@
+﻿namespace SOLIDHomework.Core.Calculators
+{
+    public interface IItemCalculator
+    {
+        decimal CalculateItemTotal(OrderItem orderItem);
+    }
+}
