+++ SOLIDHomework.Core/Calculators/ITaxCalculator.cs
@@ -1,0 +1,7 @@
+﻿namespace SOLIDHomework.Core.Calculators
+{
+    public interface ITaxCalculator
+    {
+        decimal CalculateTax(decimal total);
+    }
+}
