+++ SOLIDHomework.Core/Calculators/ItemCalculator.cs
@@ -1,0 +1,30 @@
+﻿namespace SOLIDHomework.Core.Calculators
+{
+    public class ItemCalculator : IItemCalculator
+    {
+        private IDiscountCalculator _discountCalculator;
+        public decimal CalculateItemTotal(OrderItem orderItem)
+        {
+            decimal itemTotal = 0;
+
+            if (orderItem.Type == "Unit")
+            {
+                decimal unitDiscount = _discountCalculator.CalculateUnitDiscount(orderItem);
+                itemTotal = orderItem.Amount * orderItem.Price * (1 - unitDiscount / 100m);
+            }
+            //when buy 4 prodcuts - get one for free!
+            else if (orderItem.Type == "Special")
+            {
+                itemTotal = orderItem.Amount * orderItem.Price;
+                int setsOfFour = orderItem.Amount / 4;
+                itemTotal -= setsOfFour * orderItem.Price;
+            }
+            else if (orderItem.Type == "Weight")
+            {
+                itemTotal = orderItem.Amount * orderItem.Price / 1000M;
+            }
+
+            return itemTotal;
+        }
+    }
+}
