+++ SOLIDHomework.Core/Exceptions/AccountBalanceMismatchException.cs
@@ -1,0 +1,8 @@
+﻿using System;
+
+namespace SOLIDHomework.Core.Exceptions
+{
+    public class AccountBalanceMismatchException : Exception
+    {
+    }
+}
