+++ SOLIDHomework.Core/Exceptions/OrderException.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace SOLIDHomework.Core.Exceptions
+{
+    public class OrderException : Exception
+    {
+        public OrderException(string message, Exception innerException)
+            : base(message, innerException)
+        {
+        }
+    }
+}
