+++ SOLIDHomework.Core/SOLIDHomework.Core.csproj
@@ -41,24 +41,49 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Exceptions\AccountBalanceMismatchException.cs" />
+    <Compile Include="Payment\PaymentMethod\CreditCardPayment.cs" />
+    <Compile Include="Calculators\DiscountCalculator.cs" />
+    <Compile Include="Calculators\IDiscountCalculator.cs" />
+    <Compile Include="Calculators\IItemCalculator.cs" />
+    <Compile Include="Payment\PaymentMethod\IPaymentService.cs" />
+    <Compile Include="Services\IInventory.cs" />
+    <Compile Include="Services\ILogger.cs" />
+    <Compile Include="Services\INotificationService.cs" />
+    <Compile Include="Services\Inventory.cs" />
+    <Compile Include="Calculators\ITaxCalculator.cs" />
+    <Compile Include="Calculators\ItemCalculator.cs" />
     <Compile Include="Model\Account.cs" />
     <Compile Include="Model\CreditCart.cs" />
     <Compile Include="Model\PaymentDetails.cs" />
-    <Compile Include="Model\PaymentMethod.cs" />
+    <Compile Include="Enums\PaymentMethod.cs" />
     <Compile Include="Model\OrderItem.cs" />
-    <Compile Include="OrderService.cs" />
-    <Compile Include="Payment\IPaymentBase.cs" />
-    <Compile Include="Payment\PaymentFactory.cs" />
-    <Compile Include="Payment\PaymentServiceType.cs" />
-    <Compile Include="Payment\PayPalPayment.cs" />
-    <Compile Include="Payment\PayPalWebService.cs" />
-    <Compile Include="Payment\WorldPayPayment.cs" />
-    <Compile Include="Payment\WorldPayWebService.cs" />
+    <Compile Include="Services\IUserService.cs" />
+    <Compile Include="Services\MyLogger.cs" />
+    <Compile Include="Services\IInventoryService.cs" />
+    <Compile Include="Services\IShoppingCartService.cs" />
+    <Compile Include="Services\NotificationService.cs" />
+    <Compile Include="Payment\PaymentMethod\OnlineOrderPayment.cs" />
+    <Compile Include="Exceptions\OrderException.cs" />
+    <Compile Include="Services\OrderService.cs" />
+    <Compile Include="Payment\PaymentMethod\PaymentMethodBase.cs" />
+    <Compile Include="Payment\PaymentMethod\PaymentMethodFactory.cs" />
+    <Compile Include="Payment\PaymentMethod\PaymentService.cs" />
+    <Compile Include="Payment\PaymentType\IPaymentBase.cs" />
+    <Compile Include="Payment\PaymentType\PaymentFactory.cs" />
+    <Compile Include="Enums\PaymentServiceType.cs" />
+    <Compile Include="Payment\PaymentType\PayPalPayment.cs" />
+    <Compile Include="Payment\PaymentType\PayPalWebService.cs" />
+    <Compile Include="Payment\PaymentType\WorldPayPayment.cs" />
+    <Compile Include="Payment\PaymentType\WorldPayWebService.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Services\InsufficientInventoryException.cs" />
+    <Compile Include="Exceptions\InsufficientInventoryException.cs" />
     <Compile Include="Services\InventoryService.cs" />
     <Compile Include="Services\UserService.cs" />
-    <Compile Include="ShoppingCart.cs" />
+    <Compile Include="Services\ShoppingCartService.cs" />
+    <Compile Include="Calculators\TaxCalculateFactory.cs" />
+    <Compile Include="Calculators\TaxCalculator.cs" />
+    <Compile Include="Calculators\USTaxCalculator.cs" />
   </ItemGroup>
   <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
