+++ SOLIDHomework.Core/Services/IInventory.cs
@@ -1,0 +1,7 @@
+﻿namespace SOLIDHomework.Core.Services
+{
+    public interface IInventory
+    {
+        void ReserveInventory(IShoppingCartService cart);
+    }
+}
