+++ SOLIDHomework.Core/Services/INotificationService.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace SOLIDHomework.Core.Services
+{
+    public interface INotificationService
+    {
+        void NotifyCustomer(string username);
+    }
+}
