+++ SOLIDHomework.Core/Services/IShoppingCartService.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace SOLIDHomework.Core.Services
+{
+    public interface IShoppingCartService
+    {
+        string Country { get; }
+        IEnumerable<OrderItem> OrderItems { get; }
+        void Add(OrderItem orderItem);
+        decimal TotalAmount();
+    }
+}
