+++ SOLIDHomework.Core/Services/IUserService.cs
@@ -1,0 +1,14 @@
+﻿using SOLIDHomework.Core.Model;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace SOLIDHomework.Core.Services
+{
+    public interface IUserService
+    {
+        Account GetByUsername(string username);
+    }
+}
