+++ SOLIDHomework.Core/Services/InventoryService.cs
@@ -1,6 +1,6 @@
 ﻿namespace SOLIDHomework.Core.Services
 {
-    public class InventoryService
+    public class InventoryService: IInventoryService
     {
         // that is Database-based service 
         public void Reserve(string identifier, int quantity)
