+++ SOLIDHomework.Core/Services/MyLogger.cs
@@ -1,0 +1,25 @@
+﻿using SOLIDHomework.Core.Services;
+using System.Configuration;
+using System.IO;
+
+namespace SOLIDHomework.Core
+{
+    public class MyLogger: ILogger
+    {
+        private readonly string filePath;
+        public MyLogger()
+        {
+            filePath = ConfigurationManager.AppSettings["logPath"];
+        }
+        public void Write(string text)
+        {
+            using (Stream file = File.OpenWrite(filePath))
+            {
+                using (StreamWriter writer = new StreamWriter(file))
+                {
+                    writer.WriteLine(text);
+                }
+            }
+        }
+    }
+}
